glog: store format string as data argument, remove duplicate error passing
The FormatStringArg will allow glog backends to receive the
raw first argument to an Errorf call, which can then be used
to more accurately determine an error type.

Removes duplicate code that passes error arguments as
ErrorArgs -- this is now performed in filterData, so
this special code in printfWithDepth is not needed anymore.

J=SRE-3390
TEST=auto
  Added test for format string and error arg behavior